flip-api/pyproject.toml
# Copyright (c) 2026 Guy's and St Thomas' NHS Foundation Trust & King's College London
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#     http://www.apache.org/licenses/LICENSE-2.0
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#

[project]
name = "flip-api"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "R. Garcia-Dias", email = "rafaelagd@gmail.com" },
    { name = "A. Triay Bagur", email = "alexandre.triay_bagur@kcl.ac.uk" },
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.27",
    "boto3>=1.34.49",
    "pydantic>=2.6.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "asyncpg>=0.30.0",
    "sqlmodel>=0.0.24",
    "boto3-stubs>=1.37.22",
    "types-python-jose>=3.4.0.20250224",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "sqlparse>=0.5.3",
    "pydantic-settings>=2.9.1",
    "apscheduler>=3.11.0",
]

[project.scripts]
flip-api = "flip_api:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "coverage>=7.7.1",
    "factory-boy>=3.3.3",
    "ipython>=9.0.2",
    "mypy>=1.15.0",
    "pytest-cov>=6.0.0",
    "pytest-factoryboy>=2.7.0",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "ruff>=0.11.0",
    "types-requests>=2.32.0.20250328",
    "pytest-asyncio>=0.26.0",
    "ipykernel>=6.29.5",
    "thefuzz>=0.22.1",
    "ollama>=0.4.8",
    "autopep8>=2.3.2",
    "libcst>=1.7.0",
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "pragma: no cover",
    "pass",
    "continue",
    "except Exception as e:",
    "except Exception:",
    "try:",
    "class .*\\(.*Enum.*\\):",
    "class .*\\(.*TypedDict.*\\):",
]
omit = ["*.venv/*", "*/tests/*", "*/__init__.py", "*/translation/*", "*/migrations/*"]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
addopts = []
filterwarnings = ["ignore::DeprecationWarning", "ignore::FutureWarning"]

[tool.pytest.ini_options_debug]
python_files = ["test_*.py", "*_test.py"]
filterwarnings = ["ignore::DeprecationWarning", "ignore::FutureWarning"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PT', 'UP006', 'UP007', 'UP035', 'UP045']
exclude = [
    "src/flip_api/translation/*"
]

[tool.mypy]
ignore_missing_imports = true
exclude = [
    "src/flip_api/translation/*",
]
